refactor: provider dto 패키지명을 model로 변경
