+++ src/main/java/shop/genieus/study/commons/provider/AttendanceProvider.java
@@ -3,7 +3,7 @@ package shop.genieus.study.commons.provider;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
-import shop.genieus.study.commons.provider.dto.AttendanceInfo;
+import shop.genieus.study.commons.provider.model.AttendanceInfo;
 
 public interface AttendanceProvider {
   boolean existsByUserIdAndDate(Long userId, LocalDate date);
