+++ src/main/java/shop/genieus/study/commons/provider/model/UserInfo.java
@@ -1,5 +1,4 @@
-package shop.genieus.study.commons.provider.dto;
-
+package shop.genieus.study.commons.provider.model;
 
 import java.time.LocalTime;
