+++ src/main/java/shop/genieus/study/domains/attendance/application/AttendanceQueryService.java
@@ -12,8 +12,8 @@ import org.springframework.transaction.annotation.Transactional;
 import shop.genieus.study.commons.provider.AttendanceProvider;
 import shop.genieus.study.commons.provider.DateTimeProvider;
 import shop.genieus.study.commons.provider.UserProvider;
-import shop.genieus.study.commons.provider.dto.AttendanceInfo;
-import shop.genieus.study.commons.provider.dto.UserSettingHistoryInfo;
+import shop.genieus.study.commons.provider.model.AttendanceInfo;
+import shop.genieus.study.commons.provider.model.UserSettingHistoryInfo;
 import shop.genieus.study.domains.attendance.application.dto.info.GetAttendanceInfo;
 import shop.genieus.study.domains.attendance.application.dto.result.AttendanceResult;
 import shop.genieus.study.domains.attendance.application.exception.AttendanceNotFoundException;
