+++ src/main/java/shop/genieus/study/domains/statistics/application/assembler/DailyStatisticsAssembler.java
@@ -6,8 +6,8 @@ import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import shop.genieus.study.commons.provider.dto.AttendanceInfo;
-import shop.genieus.study.commons.provider.dto.StampHistoryInfo;
+import shop.genieus.study.commons.provider.model.AttendanceInfo;
+import shop.genieus.study.commons.provider.model.StampHistoryInfo;
 import shop.genieus.study.domains.statistics.domain.vo.UserStatistics;
 
 @Slf4j
