+++ src/main/java/shop/genieus/study/domains/user/application/UserQueryService.java
@@ -8,8 +8,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.genieus.study.commons.provider.UserProvider;
-import shop.genieus.study.commons.provider.dto.UserInfo;
-import shop.genieus.study.commons.provider.dto.UserSettingHistoryInfo;
+import shop.genieus.study.commons.provider.model.UserInfo;
+import shop.genieus.study.commons.provider.model.UserSettingHistoryInfo;
 import shop.genieus.study.domains.user.application.dto.result.UserInfoResult;
 import shop.genieus.study.domains.user.application.exception.UserNotFoundException;
 import shop.genieus.study.domains.user.application.mapper.UserMapper;
