+++ src/main/java/shop/genieus/study/domains/user/application/mapper/UserMapper.java
@@ -3,8 +3,8 @@ package shop.genieus.study.domains.user.application.mapper;
 import java.util.Optional;
 import java.util.function.Function;
 import org.springframework.stereotype.Component;
-import shop.genieus.study.commons.provider.dto.UserInfo;
-import shop.genieus.study.commons.provider.dto.UserSettingHistoryInfo;
+import shop.genieus.study.commons.provider.model.UserInfo;
+import shop.genieus.study.commons.provider.model.UserSettingHistoryInfo;
 import shop.genieus.study.domains.user.domain.entity.User;
 import shop.genieus.study.domains.user.domain.entity.UserSettingHistory;
 import shop.genieus.study.domains.user.domain.vo.Email;
