+++ src/main/java/shop/genieus/study/domains/user/application/repository/UserCacheRepository.java
@@ -1,7 +1,7 @@
 package shop.genieus.study.domains.user.application.repository;
 
 import java.util.List;
-import shop.genieus.study.commons.provider.dto.UserInfo;
+import shop.genieus.study.commons.provider.model.UserInfo;
 
 public interface UserCacheRepository {
   List<UserInfo> findAllParticipatingUsers();
